chore: deprecate select & tag
BREAKING CHANGE: remove deprecated select & tag and move new implement from lab to core
