+++ package.json
@@ -58,7 +58,6 @@
     "@types/ramda": "^0.26.34",
     "@types/react": "16.9.11",
     "@types/react-dom": "^16.9.4",
-    "@types/react-select": "^2.0.19",
     "@types/styled-components": "4.4.0",
     "@types/styled-system": "^4.2.2",
     "@types/yup": "^0.26.24",
