+++ packages/tailor-ui-formik/.size-snapshot.json
@@ -1,7 +1,7 @@
 {
   "lib/index.cjs.js": {
-    "bundled": 7991,
-    "minified": 3895,
-    "gzipped": 1080
+    "bundled": 7934,
+    "minified": 3844,
+    "gzipped": 1093
   }
 }
