+++ packages/tailor-ui-lab/package.json
@@ -28,14 +28,7 @@
   "typings": "lib/index.d.ts",
   "dependencies": {
     "@tailor-ui/utils": "^0.2.9",
-    "downshift": "3.2.12",
-    "fuzzaldrin-plus": "^0.6.0",
-    "lodash.debounce": "^4.0.8",
     "polished": "^3.4.2",
-    "ramda": "^0.26.1",
-    "react-icons": "^3.8.0",
-    "react-tiny-virtual-list": "^2.2.0",
-    "styled-system": "^4.2.4",
     "tailor-ui": "^0.25.5"
   },
   "peerDependencies": {
