--- packages/tailor-ui-lab/src/Select/index.ts
@@ -1,4 +1,0 @@
-import { Option as _Option } from './SelectOptions';
-
-export * from './Select';
-export type Option = _Option;
