--- packages/tailor-ui-lab/src/Tag/index.ts
@@ -1,1 +1,0 @@
-export * from './Tag';
