+++ packages/tailor-ui-lab/src/__tests__/index.spec.ts
@@ -1,10 +1,7 @@
-import { AutoSizeInput, Select, Tabs, Tag } from '..';
+import { Tabs } from '..';
 
 describe('index', () => {
   it('should export all components', () => {
-    expect(AutoSizeInput).toBeDefined();
-    expect(Select).toBeDefined();
     expect(Tabs).toBeDefined();
-    expect(Tag).toBeDefined();
   });
 });
