+++ packages/tailor-ui-lab/src/index.ts
@@ -1,4 +1,1 @@
 export * from './Tabs';
-export * from './Select';
-export * from './Tag';
-export * from './AutoSizeInput';
