+++ packages/tailor-ui/.size-snapshot.json
@@ -1,7 +1,7 @@
 {
   "lib/index.cjs.js": {
-    "bundled": 415385,
-    "minified": 288341,
-    "gzipped": 131970
+    "bundled": 449773,
+    "minified": 304504,
+    "gzipped": 136128
   }
 }
