+++ packages/tailor-ui/src/Select/index.ts
@@ -1,1 +1,4 @@
+import { Option as _Option } from './SelectOptions';
+
 export * from './Select';
+export type Option = _Option;
