+++ packages/tailor-ui/src/Tag/Tag.tsx
@@ -1,116 +1,173 @@
-import React, { FC, useEffect, useState } from 'react';
+import React, {
+  FC,
+  FocusEvent,
+  HTMLAttributes,
+  KeyboardEvent,
+  MouseEventHandler,
+  ReactNode,
+  useCallback,
+  useState,
+} from 'react';
 import styled from 'styled-components';
 import { MdClose } from 'react-icons/md';
-import { animated, config, useSpring } from 'react-spring';
-import { omit } from 'ramda';
+import { animated, useSpring } from 'react-spring';
+
+import { useMeasure, usePrevious } from '@tailor-ui/hooks';
 
 import { Icon } from '../Icon';
 
-const StyledTag = styled.div`
-  display: inline-flex;
-  align-items: center;
-  height: ${p => p.theme.heights.sm};
-  padding: 0 ${p => p.theme.paddings.xs};
-  border: ${p => p.theme.borders.base};
-  border-radius: ${p => p.theme.radii.base};
-  border-color: ${p => p.theme.colors.gray400};
-  font-size: ${p => p.theme.fontSizes.sm};
-  white-space: nowrap;
-  cursor: pointer;
-  ${p => p.theme.transition};
-
-  &:hover {
-    border-color: ${p => p.theme.colors.gray700};
-  }
-`;
-
-const CloseIcon = styled(Icon)`
-  &:hover svg {
-    fill: ${p => p.theme.colors.gray700};
-  }
-`;
+import { StyledTag, StyledTagInput, StyledTagPrefix } from './styles';
 
 const StyledTagWrapper = styled.div`
   display: inline-flex;
-  height: ${p => p.theme.heights.sm};
-
-  & + & {
-    margin-left: 8px;
-  }
+  overflow-x: hidden;
 `;
 
 const AnimatedStyledTagWrapper = animated(StyledTagWrapper);
 
-interface ClosableTagProps {
-  onClosed?: () => void;
-}
+const StyledCloseIcon = styled(Icon)`
+  & svg {
+    fill: ${p => p.theme.colors.primaryLight2};
+  }
 
-const ClosableTag: FC<ClosableTagProps> = ({
-  children,
-  onClosed = () => {},
-  ...props
-}) => {
-  const [on, setOn] = useState(true);
+  &:hover svg {
+    fill: ${p => p.theme.colors.primaryLight};
+  }
+`;
 
-  const styles = useSpring({
-    onRest: ({ width }: any) => {
-      if (width === 0 && onClosed) {
-        onClosed();
-      }
-    },
-    immediate: on,
-    opacity: on ? 1 : 0,
-    width: on ? 'auto' : 0,
-    marginLeft: on ? 8 : 0,
-    config: config.stiff,
-  });
-
-  return (
-    <AnimatedStyledTagWrapper style={styles}>
-      <div style={{ display: 'inline-flex' }}>
-        <StyledTag {...props}>
-          {children}
-          <CloseIcon
-            size="16"
-            ml="1"
-            fill="gray400"
-            cursor="pointer"
-            onClick={() => setOn(!on)}
-            type={MdClose}
-          />
-        </StyledTag>
-      </div>
-    </AnimatedStyledTagWrapper>
-  );
-};
-
-const BaseTag: FC = ({ children, ...props }) => (
-  <StyledTagWrapper>
-    <StyledTag {...omit(['onClosed'], props)}>{children}</StyledTag>
-  </StyledTagWrapper>
+const CloseIcon: FC<{
+  onClick: MouseEventHandler;
+  invalid: boolean;
+}> = ({ onClick, invalid }) => (
+  <StyledCloseIcon
+    size="16"
+    ml="1"
+    cursor="pointer"
+    fill={invalid ? 'danger' : undefined}
+    onClick={onClick}
+    type={MdClose}
+  />
 );
 
-export interface TagProps {
+export interface TagProps
+  extends Omit<HTMLAttributes<HTMLDivElement>, 'prefix' | 'onChange'> {
   /**
    * Whether the Tag can be closed
    */
   closable?: boolean;
+  editable?: boolean;
   /**
    * Callback executed when close animation is completed
    */
   onClosed?: () => void;
+  canClose?: () => boolean | (() => Promise<boolean>);
+  onChange?: (previousValue: string, value: string) => void;
+  children?: string;
+  initialEditing?: boolean;
+  invalid?: boolean;
+  prefix?: ReactNode;
 }
 
-const Tag: FC<TagProps> = ({ closable = false, children, ...props }) => {
-  const RenderComponent = closable ? ClosableTag : BaseTag;
+const Tag: FC<TagProps> = ({
+  children = '',
+  editable = false,
+  initialEditing = false,
+  closable,
+  invalid = false,
+  onClosed,
+  canClose,
+  prefix,
+  onChange,
+  onClick,
+  ...otherProps
+}) => {
+  const [on, setOn] = useState(true);
+  const [editing, setEditing] = useState(initialEditing);
+  const previous = usePrevious(on);
+  const [bind, { width }] = useMeasure();
 
-  useEffect(() => {
-    console.warn(
-      '[tailor-ui: Tag] Tag component is deprecated. Please use the @tailor-ui/lab version instead.'
-    );
-  }, []);
+  const props = useSpring({
+    opacity: on ? 1 : 0,
+    marginRight: on ? 4 : 0,
+    width: on ? width : 0,
+    onRest: ({ width: restWidth }: { width: number }) => {
+      if (!on && restWidth === 0 && onClosed) {
+        onClosed();
+      }
+    },
+  });
 
-  return <RenderComponent {...props}>{children}</RenderComponent>;
+  const handleUpdate = useCallback(
+    (event: FocusEvent<HTMLInputElement> | KeyboardEvent<HTMLInputElement>) => {
+      const { value } = event.currentTarget;
+
+      setEditing(false);
+
+      if (onChange && children !== value) {
+        onChange(children, value);
+      }
+    },
+    [children, onChange]
+  );
+
+  return (
+    <AnimatedStyledTagWrapper
+      style={{ ...props, width: on && previous === on ? 'auto' : props.width }}
+      {...bind}
+    >
+      <StyledTag
+        editable={editable}
+        clickable={Boolean(onClick)}
+        invalid={invalid}
+        onClick={event => {
+          if (editable) {
+            setEditing(true);
+          }
+          if (onClick) {
+            onClick(event);
+          }
+        }}
+        {...otherProps}
+      >
+        {prefix && (
+          <StyledTagPrefix invalid={invalid}>{prefix}</StyledTagPrefix>
+        )}
+        {editing ? (
+          <StyledTagInput
+            autoFocus
+            defaultValue={children}
+            onBlur={handleUpdate}
+            onKeyPress={event => {
+              if (event.key === 'Enter') {
+                handleUpdate(event);
+              }
+            }}
+            onKeyDown={event => {
+              if (event.key === 'Escape') {
+                setEditing(false);
+              }
+            }}
+          />
+        ) : (
+          children
+        )}
+        {closable && (
+          <CloseIcon
+            invalid={invalid}
+            onClick={async event => {
+              event.stopPropagation();
+
+              if (canClose && !(await canClose())) {
+                return;
+              }
+
+              setOn(!on);
+            }}
+          />
+        )}
+      </StyledTag>
+    </AnimatedStyledTagWrapper>
+  );
 };
 
 export { Tag };
