+++ packages/tailor-ui/src/Tag/__tests__/Tag.spec.tsx
@@ -5,27 +5,50 @@ import { fireEvent, render, wait } from 'test/test-utils';
 import { Tag } from '../Tag';
 
 describe('Tag', () => {
-  it('should render tags correctly', () => {
-    const { queryByText } = render(
-      <div>
-        <Tag>Tag A</Tag>
-        <Tag>Tag B</Tag>
-      </div>
+  it('should render tag', () => {
+    const { container } = render(<Tag>Tailor UI</Tag>);
+
+    expect(container).toMatchSnapshot();
+  });
+
+  it('should render tag with prefix', () => {
+    const { container } = render(<Tag prefix="1">Tailor UI</Tag>);
+
+    expect(container).toMatchSnapshot();
+  });
+
+  it('should render invalid style tag', () => {
+    const { container } = render(<Tag invalid>Tailor UI</Tag>);
+
+    expect(container).toMatchSnapshot();
+  });
+
+  it('should render editable tag and call onChange', async () => {
+    const onChange = jest.fn();
+    const { container, getByText } = render(
+      <Tag editable onChange={onChange}>
+        Foo
+      </Tag>
     );
 
-    const tagA = queryByText('Tag A');
-    const tagB = queryByText('Tag B');
+    const tagText = getByText('Foo');
+    fireEvent.click(tagText);
 
-    expect(tagA).toBeInTheDocument();
-    expect(tagB).toBeInTheDocument();
+    const input = container.querySelector('input') as HTMLInputElement;
+
+    fireEvent.change(input, { target: { value: 'Foo Bar' } });
+    expect(input.value).toBe('Foo Bar');
+
+    fireEvent.keyPress(input, { key: 'Enter', code: 13, charCode: 13 });
+    expect(onChange).toBeCalledWith('Foo', 'Foo Bar');
   });
 
   it('should render closable tag', async () => {
     const { container, queryByText } = render(<Tag closable>Tag A</Tag>);
 
-    const closeIcon = container.querySelector('i') as HTMLElement;
+    const closeIcon = container.querySelector('i');
 
-    fireEvent.click(closeIcon);
+    fireEvent.click(closeIcon as HTMLElement);
 
     const tagA = queryByText('Tag A');
 
@@ -43,13 +66,51 @@ describe('Tag', () => {
       </Tag>
     );
 
-    const closeIcon = container.querySelector('i') as HTMLElement;
+    const closeIcon = container.querySelector('i');
 
-    fireEvent.click(closeIcon);
+    fireEvent.click(closeIcon as HTMLElement);
 
     await wait(() => expect(onClosed).toBeCalled());
 
     const tagA = queryByText('Tag A');
     expect(tagA).not.toBeVisible();
+  });
+
+  it('should call canClose when close icon is clicked and still visbile', () => {
+    const canClose = jest.fn().mockResolvedValue(false);
+
+    const { container, queryByText } = render(
+      <Tag closable canClose={canClose}>
+        Tailor UI
+      </Tag>
+    );
+
+    const closeIcon = container.querySelector('i');
+
+    fireEvent.click(closeIcon as HTMLElement);
+
+    const tag = queryByText('Tailor UI');
+
+    expect(canClose).toBeCalled();
+    expect(tag).toBeVisible();
+  });
+
+  it('should call canClose when close icon is clicked and close correct', async () => {
+    const canClose = jest.fn().mockResolvedValue(true);
+
+    const { container, queryByText } = render(
+      <Tag closable canClose={canClose}>
+        Tailor UI
+      </Tag>
+    );
+
+    const closeIcon = container.querySelector('i');
+
+    fireEvent.click(closeIcon as HTMLElement);
+
+    const tag = queryByText('Tailor UI');
+
+    expect(canClose).toBeCalled();
+    await (() => expect(tag).not.toBeVisible());
   });
 });
